requirements.txt
ffmpeg-python==0.2.0
pydub==0.25.1
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
faster-whisper==1.0.3
librosa==0.10.2.post1
langdetect==1.0.9
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
transformers==4.41.2
sentencepiece==0.2.0
opencv-python==4.10.0.84
mediapipe==0.10.14
onnxruntime==1.22.1
